requirements.txt
texar-pytorch
pytorch
tqdm